updated splitted unhappy path scenarios
