+++ src/test/scala/uk/gov/hmrc/test/api/specs/TfcpLinkEndpointsUnhappyPath.scala
@@ -21,9 +21,9 @@ import uk.gov.hmrc.test.api.models.User._
 
 class TfcpLinkEndpointsUnhappyPath extends BaseSpec with CommonSpec with HttpClient {
 
-  Feature("Connect to TFCP API endpoints unhappy path") {
-    var consignorToken = givenGetToken(aaResp.outboundChildPaymentRef, 250, "Individual")
-    Scenario(s"Endpoints with payload missing correlation id") {
+  Feature("Connect to TFCP API Link endpoints unhappy path") {
+    val consignorToken = givenGetToken(aaResp.outboundChildPaymentRef, 250, "Individual")
+    Scenario(s"Link endpoints with payload missing correlation id") {
       var response =
         tfcLinkWithoutCorrelationId(
           consignorToken,
@@ -35,56 +35,16 @@ class TfcpLinkEndpointsUnhappyPath extends BaseSpec with CommonSpec with HttpCli
       thenValidateResponseCode(response, 400)
       checkJsonValue(response, "errorCode", "ETFC1")
       checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response =
-        tfcBalanceWithoutCorrelationId(consignorToken, eppUniqueCusId, eppRegRef, aaResp.outboundChildPaymentRef)
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "ETFC1")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcPaymentWithoutCorrelationId(
-        consignorToken,
-        eppUniqueCusId,
-        eppRegRef,
-        aaResp.outboundChildPaymentRef,
-        paymentAmount,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "ETFC1")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
     }
-    Scenario(s"Endpoints payload with an empty correlation id") {
-      var response =
+    Scenario(s"Link endpoints payload with an empty correlation id") {
+      val response =
         tfcLink(consignorToken, "", eppUniqueCusId, eppRegRef, aaResp.outboundChildPaymentRef, childDOB)
       thenValidateResponseCode(response, 400)
       checkJsonValue(response, "errorCode", "ETFC1")
       checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcBalance(consignorToken, "", eppUniqueCusId, eppRegRef, aaResp.outboundChildPaymentRef)
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "ETFC1")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcPayment(
-        consignorToken,
-        "",
-        eppUniqueCusId,
-        eppRegRef,
-        aaResp.outboundChildPaymentRef,
-        paymentAmount,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "ETFC1")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
     }
-    Scenario(s"Endpoints payload with an correlation id as int") {
-      var response =
+    Scenario(s"Link endpoints payload with an correlation id as int") {
+      val response =
         tfcLinkInvalidDataTypeCorrelationID(
           consignorToken,
           123,
@@ -96,90 +56,23 @@ class TfcpLinkEndpointsUnhappyPath extends BaseSpec with CommonSpec with HttpCli
       thenValidateResponseCode(response, 400)
       checkJsonValue(response, "errorCode", "ETFC1")
       checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcBalanceInvalidDataTypeCorrelationID(
-        consignorToken,
-        123,
-        eppUniqueCusId,
-        eppRegRef,
-        aaResp.outboundChildPaymentRef
-      )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "ETFC1")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcPaymentInvalidDataTypeCorrelationID(
-        consignorToken,
-        123,
-        eppUniqueCusId,
-        eppRegRef,
-        aaResp.outboundChildPaymentRef,
-        paymentAmount,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "ETFC1")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
     }
-    Scenario(s"Endpoints payload with an invalid correlation id") {
-      var response =
+    Scenario(s"Link endpoints payload with an invalid correlation id") {
+      val response =
         tfcLink(consignorToken, "1234", eppUniqueCusId, eppRegRef, aaResp.outboundChildPaymentRef, childDOB)
       thenValidateResponseCode(response, 400)
       checkJsonValue(response, "errorCode", "ETFC1")
       checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcBalance(consignorToken, "1234", eppUniqueCusId, eppRegRef, aaResp.outboundChildPaymentRef)
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "ETFC1")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcPayment(
-        consignorToken,
-        "1234",
-        eppUniqueCusId,
-        eppRegRef,
-        aaResp.outboundChildPaymentRef,
-        paymentAmount,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "ETFC1")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
     }
-    Scenario(s"Endpoints with a payload with an empty EPP unique customer ID") {
-      var response =
+    Scenario(s"Link endpoints with a payload with an empty EPP unique customer ID") {
+      val response =
         tfcLink(consignorToken, correlationId, "", eppRegRef, aaResp.outboundChildPaymentRef, childDOB)
       thenValidateResponseCode(response, 400)
       checkJsonValue(response, "errorCode", "E0000")
       checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcBalance(consignorToken, correlationId, "", eppRegRef, aaResp.outboundChildPaymentRef)
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0000")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcPayment(
-        consignorToken,
-        correlationId,
-        "",
-        eppRegRef,
-        aaResp.outboundChildPaymentRef,
-        paymentAmount,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0000")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
     }
-    Scenario(s"Endpoints with a payload containing a different data type of EPP unique customer ID") {
-      var response =
+    Scenario(s"Link endpoints with a payload containing a different data type of EPP unique customer ID") {
+      val response =
         tfcLinkInvalidDataTypeEPPUniqueCusId(
           consignorToken,
           correlationId,
@@ -191,36 +84,9 @@ class TfcpLinkEndpointsUnhappyPath extends BaseSpec with CommonSpec with HttpCli
       thenValidateResponseCode(response, 400)
       checkJsonValue(response, "errorCode", "E0000")
       checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcBalanceInvalidDataTypeEPPUniqueCusId(
-        consignorToken,
-        correlationId,
-        123,
-        eppRegRef,
-        aaResp.outboundChildPaymentRef
-      )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0000")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcPaymentInvalidDataTypeEPPUniqueCusId(
-        consignorToken,
-        correlationId,
-        123,
-        eppRegRef,
-        aaResp.outboundChildPaymentRef,
-        paymentAmount,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0000")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
     }
-    Scenario(s"Endpoints with a payload containing a invalid json field of EPP unique customer ID") {
-      var response =
+    Scenario(s"Link endpoints with a payload containing a invalid json field of EPP unique customer ID") {
+      val response =
         tfcLinkInvalidFieldEPPUniqueCusId(
           consignorToken,
           correlationId,
@@ -232,90 +98,23 @@ class TfcpLinkEndpointsUnhappyPath extends BaseSpec with CommonSpec with HttpCli
       thenValidateResponseCode(response, 400)
       checkJsonValue(response, "errorCode", "E0000")
       checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcBalanceInvalidFieldEPPUniqueCusId(
-        consignorToken,
-        correlationId,
-        eppUniqueCusId,
-        eppRegRef,
-        aaResp.outboundChildPaymentRef
-      )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0000")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcPaymentInvalidFieldEPPUniqueCusId(
-        consignorToken,
-        correlationId,
-        eppUniqueCusId,
-        eppRegRef,
-        aaResp.outboundChildPaymentRef,
-        paymentAmount,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0000")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
     }
-    Scenario(s"Endpoints with a payload with an missing EPP unique customer ID") {
-      var response =
+    Scenario(s"Link endpoints with a payload with an missing EPP unique customer ID") {
+      val response =
         tfcLinkWithoutEPPCusId(consignorToken, correlationId, eppRegRef, aaResp.outboundChildPaymentRef, childDOB)
       thenValidateResponseCode(response, 400)
       checkJsonValue(response, "errorCode", "E0004")
       checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcBalanceWithoutEppRegRef(consignorToken, correlationId, eppRegRef, aaResp.outboundChildPaymentRef)
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0004")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcPaymentWithoutEppCusId(
-        consignorToken,
-        correlationId,
-        eppRegRef,
-        aaResp.outboundChildPaymentRef,
-        paymentAmount,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0004")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
     }
-    Scenario(s"Payload with an empty EPP registration reference") {
-      var response =
+    Scenario(s"Link endpoint payload with an empty EPP registration reference") {
+      val response =
         tfcLink(consignorToken, correlationId, eppUniqueCusId, "", aaResp.outboundChildPaymentRef, childDOB)
       thenValidateResponseCode(response, 400)
       checkJsonValue(response, "errorCode", "E0000")
       checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcBalance(consignorToken, correlationId, eppUniqueCusId, "", aaResp.outboundChildPaymentRef)
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0000")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcPayment(
-        consignorToken,
-        correlationId,
-        eppUniqueCusId,
-        "",
-        aaResp.outboundChildPaymentRef,
-        paymentAmount,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0000")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
     }
-    Scenario(s"Payload with an invalid json field of EPP registration reference") {
-      var response =
+    Scenario(s"Link endpoint payload with an invalid json field of EPP registration reference") {
+      val response =
         tfcLinkInvalidFieldEPPRegRef(
           consignorToken,
           correlationId,
@@ -327,35 +126,9 @@ class TfcpLinkEndpointsUnhappyPath extends BaseSpec with CommonSpec with HttpCli
       thenValidateResponseCode(response, 400)
       checkJsonValue(response, "errorCode", "E0000")
       checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcBalanceInvalidFieldEPPRegRef(
-        consignorToken,
-        correlationId,
-        eppUniqueCusId,
-        eppRegRef,
-        aaResp.outboundChildPaymentRef
-      )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0000")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcPaymentInvalidFieldEPPRegRef(
-        consignorToken,
-        correlationId,
-        eppUniqueCusId,
-        eppRegRef,
-        aaResp.outboundChildPaymentRef,
-        paymentAmount,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0000")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
     }
-    Scenario(s"Payload with an different data type of EPP registration reference") {
-      var response =
+    Scenario(s"Link endpoint payload with an different data type of EPP registration reference") {
+      val response =
         tfcLinkInvalidDataTypeEPPRegRef(
           consignorToken,
           correlationId,
@@ -367,35 +140,9 @@ class TfcpLinkEndpointsUnhappyPath extends BaseSpec with CommonSpec with HttpCli
       thenValidateResponseCode(response, 400)
       checkJsonValue(response, "errorCode", "E0000")
       checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcBalanceInvalidDataTypeEPPRegRef(
-        consignorToken,
-        correlationId,
-        eppUniqueCusId,
-        123,
-        aaResp.outboundChildPaymentRef
-      )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0000")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcPaymentInvalidDataTypeEPPRegRef(
-        consignorToken,
-        correlationId,
-        eppUniqueCusId,
-        123,
-        aaResp.outboundChildPaymentRef,
-        paymentAmount,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0000")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
     }
-    Scenario(s"Payload with an missing EPP registration reference") {
-      var response =
+    Scenario(s"Link endpoint payload with an missing EPP registration reference") {
+      val response =
         tfcLinkWithoutEppRegRef(
           consignorToken,
           correlationId,
@@ -406,51 +153,11 @@ class TfcpLinkEndpointsUnhappyPath extends BaseSpec with CommonSpec with HttpCli
       thenValidateResponseCode(response, 400)
       checkJsonValue(response, "errorCode", "E0002")
       checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response =
-        tfcBalanceWithoutEppRegRef(consignorToken, correlationId, eppUniqueCusId, aaResp.outboundChildPaymentRef)
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0002")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcPaymentWithouteppRegRef(
-        consignorToken,
-        correlationId,
-        eppUniqueCusId,
-        aaResp.outboundChildPaymentRef,
-        paymentAmount,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0002")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
     }
 
-    Scenario(s"Payload with an unmatched Outbound child payment reference number in the stub") {
-      var response =
+    Scenario(s"Link endpoint payload with an unmatched Outbound child payment reference number in the stub") {
+      val response =
         tfcLink(consignorToken, correlationId, eppUniqueCusId, eppRegRef, "AAaa00000TFC", childDOB)
-      thenValidateResponseCode(response, 500)
-      checkJsonValue(response, "errorCode", "E0000")
-      checkJsonValue(response, "errorDescription", "The server encountered an error and couldn't process the request")
-
-      response = tfcBalance(consignorToken, correlationId, eppUniqueCusId, eppRegRef, "AAaa00000TFC")
-      thenValidateResponseCode(response, 500)
-      checkJsonValue(response, "errorCode", "E0000")
-      checkJsonValue(response, "errorDescription", "The server encountered an error and couldn't process the request")
-
-      response = tfcPayment(
-        consignorToken,
-        correlationId,
-        eppUniqueCusId,
-        eppRegRef,
-        "AAaa00000TFC",
-        paymentAmount,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
       thenValidateResponseCode(response, 500)
       checkJsonValue(response, "errorCode", "E0000")
       checkJsonValue(response, "errorDescription", "The server encountered an error and couldn't process the request")
@@ -468,63 +175,24 @@ class TfcpLinkEndpointsUnhappyPath extends BaseSpec with CommonSpec with HttpCli
 
     scenarios.foreach { outboundPayRef =>
       Scenario(
-        s"Verify Link Balance and Payments endpoint for predefined outbound child Payment reference number: $outboundPayRef"
+        s"Verify Link endpoint for predefined outbound child Payment reference number: $outboundPayRef"
       ) {
-        var response =
+        val response =
           tfcLink(consignorToken, correlationId, eppUniqueCusId, eppRegRef, outboundPayRef, childDOB)
-        thenValidateResponseCode(response, 400)
-        checkJsonValue(response, "errorCode", "E0000")
-        checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-        response = tfcBalance(consignorToken, correlationId, eppUniqueCusId, eppRegRef, outboundPayRef)
-        thenValidateResponseCode(response, 400)
-        checkJsonValue(response, "errorCode", "E0000")
-        checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-        response = tfcPayment(
-          consignorToken,
-          correlationId,
-          eppUniqueCusId,
-          eppRegRef,
-          outboundPayRef,
-          paymentAmount,
-          ccpRegReference,
-          ccpPostcode,
-          payeeType
-        )
         thenValidateResponseCode(response, 400)
         checkJsonValue(response, "errorCode", "E0000")
         checkJsonValue(response, "errorDescription", jsonErrorDescription)
       }
     }
-    Scenario(s"Payload with an missing Outbound child payment reference number") {
-      var response =
+    Scenario(s"Link endpoint payload with an missing Outbound child payment reference number") {
+      val response =
         tfcLinkWithoutOutboundChildPayRef(consignorToken, correlationId, eppUniqueCusId, eppRegRef, childDOB)
       thenValidateResponseCode(response, 400)
       checkJsonValue(response, "errorCode", "E0001")
       checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcBalanceWithoutOutboundChildPayRef(consignorToken, correlationId, eppUniqueCusId, eppRegRef)
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0001")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcPaymentWithoutOutboundChildPayRef(
-        consignorToken,
-        correlationId,
-        eppUniqueCusId,
-        eppRegRef,
-        paymentAmount,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0001")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
     }
-    Scenario(s"Payload with an invalid data type Outbound child payment reference number") {
-      var response =
+    Scenario(s"Link endpoint payload with an invalid data type Outbound child payment reference number") {
+      val response =
         tfcLinkInvalidDataTypeOutboundChildPayRef(
           consignorToken,
           correlationId,
@@ -536,30 +204,9 @@ class TfcpLinkEndpointsUnhappyPath extends BaseSpec with CommonSpec with HttpCli
       thenValidateResponseCode(response, 400)
       checkJsonValue(response, "errorCode", "E0001")
       checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response =
-        tfcBalanceInvalidDataTypeOutboundChildPayRef(consignorToken, correlationId, eppUniqueCusId, eppRegRef, 123)
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0001")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcPaymentInvalidDataTypeOutboundChildPayRef(
-        consignorToken,
-        correlationId,
-        eppUniqueCusId,
-        eppRegRef,
-        paymentAmount,
-        123,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0001")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
     }
-    Scenario(s"Payload with an invalid Field Outbound child payment reference number") {
-      var response =
+    Scenario(s"Link endpoint payload with an invalid Field Outbound child payment reference number") {
+      val response =
         tfcLinkInvalidFieldOutboundChildPayRef(
           consignorToken,
           correlationId,
@@ -568,32 +215,6 @@ class TfcpLinkEndpointsUnhappyPath extends BaseSpec with CommonSpec with HttpCli
           aaResp.outboundChildPaymentRef,
           childDOB
         )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0001")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcBalanceInvalidFieldOutboundChildPayRef(
-        consignorToken,
-        correlationId,
-        eppUniqueCusId,
-        eppRegRef,
-        aaResp.outboundChildPaymentRef
-      )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0001")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-
-      response = tfcPaymentInvalidFieldOutboundChildPayRef(
-        consignorToken,
-        correlationId,
-        eppUniqueCusId,
-        eppRegRef,
-        aaResp.outboundChildPaymentRef,
-        paymentAmount,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
       thenValidateResponseCode(response, 400)
       checkJsonValue(response, "errorCode", "E0001")
       checkJsonValue(response, "errorDescription", jsonErrorDescription)
@@ -659,505 +280,6 @@ class TfcpLinkEndpointsUnhappyPath extends BaseSpec with CommonSpec with HttpCli
       thenValidateResponseCode(response, 400)
       checkJsonValue(response, "errorCode", "E0006")
       checkJsonValue(response, "errorDescription", jsonErrorDescription)
-    }
-    Scenario(s"Payments with a payload with an empty Payment amount") {
-      val response =
-        tfcPaymentInvalidPaymentAmount(
-          consignorToken,
-          correlationId,
-          eppUniqueCusId,
-          eppRegRef,
-          aaResp.outboundChildPaymentRef,
-          "",
-          ccpRegReference,
-          ccpPostcode,
-          payeeType
-        )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0023")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-    }
-    Scenario(s"Payments with a payload with an invalid field Payment amount") {
-      val response =
-        tfcPaymentInvalidFieldPaymentAmount(
-          consignorToken,
-          correlationId,
-          eppUniqueCusId,
-          eppRegRef,
-          aaResp.outboundChildPaymentRef,
-          paymentAmount,
-          ccpRegReference,
-          ccpPostcode,
-          payeeType
-        )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0023")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-    }
-    Scenario(s"Payments with a payload with an invalid data type Payment amount") {
-      val response =
-        tfcPaymentInvalidPaymentAmount(
-          consignorToken,
-          correlationId,
-          eppUniqueCusId,
-          eppRegRef,
-          aaResp.outboundChildPaymentRef,
-          "1234",
-          ccpRegReference,
-          ccpPostcode,
-          payeeType
-        )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0023")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-    }
-    //    Scenario(s"Payments with a payload with an Payment amount zero") {
-    //      val response =
-    //        tfcPayment(
-    //          consignorToken,
-    //          correlationId,
-    //          eppUniqueCusId,
-    //          eppRegRef,
-    //          aaResp.outboundChildPaymentRef,
-    //          0,
-    //          ccpRegReference,
-    //          ccpPostcode,
-    //          payeeType
-    //        )
-    //      thenValidateResponseCode(response, 400)
-    //      checkJsonValue(response, "errorCode", "E0023")
-    //      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-    //    }
-    //    Scenario(s"Payments with a payload with an Payment amount negative number") {
-    //      val response =
-    //        tfcPayment(
-    //          consignorToken,
-    //          correlationId,
-    //          eppUniqueCusId,
-    //          eppRegRef,
-    //          aaResp.outboundChildPaymentRef,
-    //          -123,
-    //          ccpRegReference,
-    //          ccpPostcode,
-    //          payeeType
-    //        )
-    //      thenValidateResponseCode(response, 400)
-    //      checkJsonValue(response, "errorCode", "E0023")
-    //      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-    //    }
-    Scenario(s"Payments with a payload with an missing Payment amount") {
-      val response =
-        tfcPaymentWithoutPaymentAmount(
-          consignorToken,
-          correlationId,
-          eppUniqueCusId,
-          eppRegRef,
-          aaResp.outboundChildPaymentRef,
-          ccpRegReference,
-          ccpPostcode,
-          payeeType
-        )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0008")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-    }
-    Scenario(s"Payments with a payload with an invalid ccp Reg Reference data type") {
-      val response =
-        tfcPaymentWithInvalidCCPRegReference(
-          consignorToken,
-          correlationId,
-          eppUniqueCusId,
-          eppRegRef,
-          aaResp.outboundChildPaymentRef,
-          paymentAmount,
-          123,
-          ccpPostcode,
-          payeeType
-        )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0000")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-    }
-    Scenario(s"Payments with a payload with an invalid field ccp Reg Reference") {
-      val response =
-        tfcPaymentWithInvalidFieldCCPRegReference(
-          consignorToken,
-          correlationId,
-          eppUniqueCusId,
-          eppRegRef,
-          aaResp.outboundChildPaymentRef,
-          paymentAmount,
-          ccpRegReference,
-          ccpPostcode,
-          payeeType
-        )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0000")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-    }
-    Scenario(s"Payments with a payload with an ccp Reg Reference as empty") {
-      val response =
-        tfcPayment(
-          consignorToken,
-          correlationId,
-          eppUniqueCusId,
-          eppRegRef,
-          aaResp.outboundChildPaymentRef,
-          paymentAmount,
-          "",
-          ccpPostcode,
-          payeeType
-        )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0000")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-    }
-    Scenario(s"Payments with a payload with an missing ccp Reg Reference") {
-      val response =
-        tfcPaymentWithoutCcpRegReference(
-          consignorToken,
-          correlationId,
-          eppUniqueCusId,
-          eppRegRef,
-          aaResp.outboundChildPaymentRef,
-          paymentAmount,
-          ccpPostcode,
-          payeeType
-        )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0003")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-    }
-    Scenario(s"Payments with a payload with an invalid ccp Postcode data type") {
-      val response =
-        tfcPaymentWithInvalidccpPostcode(
-          consignorToken,
-          correlationId,
-          eppUniqueCusId,
-          eppRegRef,
-          aaResp.outboundChildPaymentRef,
-          paymentAmount,
-          ccpRegReference,
-          123,
-          payeeType
-        )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0000")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-    }
-    Scenario(s"Payments with a payload with an invalid field ccp Postcode") {
-      val response =
-        tfcPaymentInvalidFieldPostcode(
-          consignorToken,
-          correlationId,
-          eppUniqueCusId,
-          eppRegRef,
-          aaResp.outboundChildPaymentRef,
-          paymentAmount,
-          ccpRegReference,
-          ccpPostcode,
-          payeeType
-        )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0000")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-    }
-    Scenario(s"Payments with a payload with an missing ccp Postcode") {
-      val response =
-        tfcPaymentWithoutCcpPostcode(
-          consignorToken,
-          correlationId,
-          eppUniqueCusId,
-          eppRegRef,
-          aaResp.outboundChildPaymentRef,
-          paymentAmount,
-          ccpRegReference,
-          payeeType
-        )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "BAD_REQUEST")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-    }
-    Scenario(s"Payments with a payload with an ccp Postcode as empty") {
-      val response =
-        tfcPayment(
-          consignorToken,
-          correlationId,
-          eppUniqueCusId,
-          eppRegRef,
-          aaResp.outboundChildPaymentRef,
-          paymentAmount,
-          ccpRegReference,
-          "",
-          payeeType
-        )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "BAD_REQUEST")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-    }
-    val ccpSscenarios =
-      List(
-        "",
-        "payeeType",
-        "EPP",
-        "12CCP",
-        "CCP12",
-        "ccp",
-        "cCp"
-      )
-
-    ccpSscenarios.foreach { ccp =>
-      Scenario(s"Verify Link Balance and Payments endpoint for predefined payee type : $ccp") {
-        val response = tfcPayment(
-          consignorToken,
-          correlationId,
-          eppUniqueCusId,
-          eppRegRef,
-          aaResp.outboundChildPaymentRef,
-          paymentAmount,
-          ccpRegReference,
-          ccpPostcode,
-          ccp
-        )
-        thenValidateResponseCode(response, 400)
-        checkJsonValue(response, "errorCode", "E0022")
-        checkJsonValue(response, "errorDescription", jsonErrorDescription)
-      }
-    }
-    Scenario(s"Payments with a payload with an missing payee Type") {
-      val response =
-        tfcPaymentWithoutPayeeType(
-          consignorToken,
-          correlationId,
-          eppUniqueCusId,
-          eppRegRef,
-          aaResp.outboundChildPaymentRef,
-          paymentAmount,
-          ccpRegReference,
-          ccpPostcode
-        )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0007")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-    }
-    Scenario(s"Payments with a payload with an invalid field payee Type") {
-      val response =
-        tfcPaymentInvalidFieldPayeeType(
-          consignorToken,
-          correlationId,
-          eppUniqueCusId,
-          eppRegRef,
-          aaResp.outboundChildPaymentRef,
-          paymentAmount,
-          ccpRegReference,
-          ccpPostcode,
-          payeeType
-        )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0007")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-    }
-    Scenario(s"Payments with a payload with an invalid data type payee Type") {
-      val response =
-        tfcPaymentInvalidDataTypePayeeType(
-          consignorToken,
-          correlationId,
-          eppUniqueCusId,
-          eppRegRef,
-          aaResp.outboundChildPaymentRef,
-          paymentAmount,
-          ccpRegReference,
-          ccpPostcode,
-          123
-        )
-      thenValidateResponseCode(response, 400)
-      checkJsonValue(response, "errorCode", "E0007")
-      checkJsonValue(response, "errorDescription", jsonErrorDescription)
-    }
-    Scenario(s"Endpoints missing authorization header") {
-      var response =
-        tfcLinkWithoutAuthorizationHeader(
-          correlationId,
-          eppUniqueCusId,
-          eppRegRef,
-          aaResp.outboundChildPaymentRef,
-          childDOB
-        )
-      thenValidateResponseCode(response, 401)
-//      checkJsonValue(response, "errorCode", "UNAUTHORISED")
-//      checkJsonValue(response, "errorDescription", "Invalid authentication credentials")
-
-      response =
-        tfcBalanceWithoutAuthorization(correlationId, eppUniqueCusId, eppRegRef, aaResp.outboundChildPaymentRef)
-      thenValidateResponseCode(response, 401)
-//      checkJsonValue(response, "errorCode", "UNAUTHORISED")
-//      checkJsonValue(response, "errorDescription", "Invalid authentication credentials")
-
-      response = tfcPaymentWithoutAuthorization(
-        correlationId,
-        eppUniqueCusId,
-        eppRegRef,
-        aaResp.outboundChildPaymentRef,
-        paymentAmount,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
-    }
-    Scenario(s"Endpoints with a token with insufficient confidence level") {
-      consignorToken = givenGetToken(aaResp.outboundChildPaymentRef, 50, "Individual")
-      var response =
-        tfcLink(consignorToken, correlationId, eppUniqueCusId, eppRegRef, aaResp.outboundChildPaymentRef, childDOB)
-      thenValidateResponseCode(response, 401)
-//      checkJsonValue(response, "errorCode", "UNAUTHORISED")
-//      checkJsonValue(response, "errorDescription", "Invalid authentication credentials")
-
-      response = tfcBalance(consignorToken, correlationId, eppUniqueCusId, eppRegRef, aaResp.outboundChildPaymentRef)
-      thenValidateResponseCode(response, 401)
-//      checkJsonValue(response, "errorCode", "UNAUTHORISED")
-//      checkJsonValue(response, "errorDescription", "Invalid authentication credentials")
-
-      response = tfcPayment(
-        consignorToken,
-        correlationId,
-        eppUniqueCusId,
-        eppRegRef,
-        aaResp.outboundChildPaymentRef,
-        paymentAmount,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
-      thenValidateResponseCode(response, 401)
-//      checkJsonValue(response, "errorCode", "UNAUTHORISED")
-//      checkJsonValue(response, "errorDescription", "Invalid authentication credentials")
-    }
-    Scenario(s"Endpoints with a token with affinity group Organisation") {
-      consignorToken = givenGetToken(aaResp.outboundChildPaymentRef, 250, "Organisation")
-      var response =
-        tfcLink(consignorToken, correlationId, eppUniqueCusId, eppRegRef, aaResp.outboundChildPaymentRef, childDOB)
-      thenValidateResponseCode(response, 401)
-//      checkJsonValue(response, "errorCode", "UNAUTHORISED")
-//      checkJsonValue(response, "errorDescription", "Invalid authentication credentials")
-
-      response = tfcBalance(consignorToken, correlationId, eppUniqueCusId, eppRegRef, aaResp.outboundChildPaymentRef)
-      thenValidateResponseCode(response, 401)
-//      checkJsonValue(response, "errorCode", "UNAUTHORISED")
-//      checkJsonValue(response, "errorDescription", "Invalid authentication credentials")
-
-      response = tfcPayment(
-        consignorToken,
-        correlationId,
-        eppUniqueCusId,
-        eppRegRef,
-        aaResp.outboundChildPaymentRef,
-        paymentAmount,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
-      thenValidateResponseCode(response, 401)
-//      checkJsonValue(response, "errorCode", "UNAUTHORISED")
-//      checkJsonValue(response, "errorDescription", "Invalid authentication credentials")
-    }
-    Scenario(s"Endpoints with a token with affinity group Agent") {
-      consignorToken = givenGetToken(aaResp.outboundChildPaymentRef, 250, "Agent")
-      var response =
-        tfcLink(consignorToken, correlationId, eppUniqueCusId, eppRegRef, aaResp.outboundChildPaymentRef, childDOB)
-      thenValidateResponseCode(response, 401)
-//      checkJsonValue(response, "errorCode", "UNAUTHORISED")
-//      checkJsonValue(response, "errorDescription", "Invalid authentication credentials")
-
-      response = tfcBalance(consignorToken, correlationId, eppUniqueCusId, eppRegRef, aaResp.outboundChildPaymentRef)
-      thenValidateResponseCode(response, 401)
-//      checkJsonValue(response, "errorCode", "UNAUTHORISED")
-//      checkJsonValue(response, "errorDescription", "Invalid authentication credentials")
-
-      response = tfcPayment(
-        consignorToken,
-        correlationId,
-        eppUniqueCusId,
-        eppRegRef,
-        aaResp.outboundChildPaymentRef,
-        paymentAmount,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
-      thenValidateResponseCode(response, 401)
-//      checkJsonValue(response, "errorCode", "UNAUTHORISED")
-//      checkJsonValue(response, "errorDescription", "Invalid authentication credentials")
-    }
-    Scenario(s"Endpoints with bearer token expired") {
-      consignorToken =
-        "Bearer BXQ3/Treo4kQCZvVcCqKPoUOMmoBVy2UTaeDDgEhL3PTJijyU/5xyYQENUp4hMYDp1T3Gze4WkHQsusa967ZIKFulM6yR9mRKsZqpQqpjcLkm3OMGi/7U4hjAhKbWEZu4dvoCxWQcOCGXI/nMQlddHdXv2ZKEdcJ8bTUTpO0WX/9KwIkeIPK/mMlBESjue4V"
-      var response =
-        tfcLink(
-          consignorToken,
-          correlationId,
-          eppUniqueCusId,
-          eppRegRef,
-          aaResp.outboundChildPaymentRef,
-          childDOB
-        )
-      thenValidateResponseCode(response, 401)
-//      checkJsonValue(response, "errorCode", "UNAUTHORISED")
-//      checkJsonValue(response, "errorDescription", "Invalid authentication credentials")
-
-      response = tfcBalance(consignorToken, correlationId, eppUniqueCusId, eppRegRef, aaResp.outboundChildPaymentRef)
-      thenValidateResponseCode(response, 401)
-//      checkJsonValue(response, "errorCode", "UNAUTHORISED")
-//      checkJsonValue(response, "errorDescription", "Invalid authentication credentials")
-
-      response = tfcPayment(
-        consignorToken,
-        correlationId,
-        eppUniqueCusId,
-        eppRegRef,
-        aaResp.outboundChildPaymentRef,
-        paymentAmount,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
-      thenValidateResponseCode(response, 401)
-//      checkJsonValue(response, "errorCode", "UNAUTHORISED")
-//      checkJsonValue(response, "errorDescription", "Invalid authentication credentials")
-    }
-    Scenario(s"Endpoints with a invalid bearer token") {
-      consignorToken = "this is invalid bearer token"
-      var response =
-        tfcLink(
-          consignorToken,
-          correlationId,
-          eppUniqueCusId,
-          eppRegRef,
-          aaResp.outboundChildPaymentRef,
-          childDOB
-        )
-      thenValidateResponseCode(response, 401)
-//      checkJsonValue(response, "errorCode", "UNAUTHORISED")
-//      checkJsonValue(response, "errorDescription", "Invalid authentication credentials")
-
-      response = tfcBalance(consignorToken, correlationId, eppUniqueCusId, eppRegRef, aaResp.outboundChildPaymentRef)
-      thenValidateResponseCode(response, 401)
-//      checkJsonValue(response, "errorCode", "UNAUTHORISED")
-//      checkJsonValue(response, "errorDescription", "Invalid authentication credentials")
-
-      response = tfcPayment(
-        consignorToken,
-        correlationId,
-        eppUniqueCusId,
-        eppRegRef,
-        aaResp.outboundChildPaymentRef,
-        paymentAmount,
-        ccpRegReference,
-        ccpPostcode,
-        payeeType
-      )
-      thenValidateResponseCode(response, 401)
-//      checkJsonValue(response, "errorCode", "UNAUTHORISED")
-//      checkJsonValue(response, "errorDescription", "Invalid authentication credentials")
     }
   }
 }
